refactor fitness
