--- src/Solver/Fitness/ICalculadoraFitness.cs
@@ -1,9 +1,0 @@
-﻿using Solver.Individuos;
-
-namespace Solver.Fitness
-{
-    internal interface ICalculadoraFitness
-    {
-        decimal CalcularFitness(Individuo individuo, InstanciaProblema instanciaProblema);
-    }
-}
